Move presim logic to its own file
