+++ sim/core/sim.go
@@ -8,7 +8,6 @@ import (
 	"time"
 
 	"github.com/wowsims/tbc/sim/core/proto"
-	googleProto "google.golang.org/protobuf/proto"
 )
 
 func debugFunc(sim *Simulation) func(string, ...interface{}) {
@@ -95,67 +94,6 @@ func (sim *Simulation) RandomFloat(label string) float64 {
 		sim.testRands[labelHash] = labelRand
 	}
 	return labelRand.Float64()
-}
-
-func (sim *Simulation) runPresims(request proto.RaidSimRequest) {
-	const numPresimIterations = 100
-
-	// Run presims if requested.
-	raidPresimOptions := make([]*PresimOptions, sim.Raid.Size())
-	remainingAgents := 0
-	for _, party := range sim.Raid.Parties {
-		for _, player := range party.Players {
-			presimOptions := player.GetPresimOptions()
-			if presimOptions != nil {
-				raidPresimOptions[player.GetCharacter().ID] = presimOptions
-				remainingAgents++
-			}
-		}
-	}
-
-	// Base presim request.
-	// Define this outside the loop so that, as Agents iteratively update their
-	// settings, we keep the most recent settings even after that Agent is
-	// done with presims.
-	presimRequest := googleProto.Clone(&request).(*proto.RaidSimRequest)
-	presimRequest.SimOptions.RandomSeed = 1
-	presimRequest.SimOptions.Debug = false
-	presimRequest.SimOptions.Iterations = numPresimIterations
-
-	for remainingAgents > 0 {
-		// ** Run a presim round. **
-
-		// Let each Agent modify their own settings.
-		for partyIdx, party := range sim.Raid.Parties {
-			partyConfig := presimRequest.Raid.Parties[partyIdx]
-			for playerIdx, player := range party.Players {
-				playerConfig := partyConfig.Players[playerIdx]
-				presimOptions := raidPresimOptions[player.GetCharacter().ID]
-				if presimOptions != nil {
-					presimOptions.SetPresimPlayerOptions(playerConfig)
-				}
-			}
-		}
-
-		// Run the presim.
-		presimResult := RunSim(*presimRequest)
-
-		// Provide each Agent with their own results.
-		for partyIdx, party := range sim.Raid.Parties {
-			partyMetrics := presimResult.RaidMetrics.Parties[partyIdx]
-			for playerIdx, player := range party.Players {
-				playerMetrics := partyMetrics.Players[playerIdx]
-				presimOptions := raidPresimOptions[player.GetCharacter().ID]
-				if presimOptions != nil {
-					done := presimOptions.OnPresimResult(*playerMetrics, numPresimIterations)
-					if done {
-						raidPresimOptions[player.GetCharacter().ID] = nil
-						remainingAgents--
-					}
-				}
-			}
-		}
-	}
 }
 
 // Reset will set sim back and erase all current state.
